+++ src/Provider/HelloWorldInformationFactoryProvider.php
@@ -8,7 +8,6 @@ use CodeHqDk\RepositoryInformation\Service\ProviderDependencyService;
 
 class HelloWorldInformationFactoryProvider implements InformationFactoryProvider
 {
-
     public function addFactory(ProviderDependencyService $provider_dependency_service): void
     {
         $provider_dependency_service->registerClassInDependencyContainer(HelloWorldInformationFactory::class);