+++ RELEASE_NOTES.md
@@ -3,5 +3,6 @@
 
 ## Version 0.0.0
 ### Changes
+* Added tests and github actions
 * Use BaseInformationBlock class
 * Initial version